AS7-6071 Split out writing an reading logic, and remove all the dead writing code
